11.4 source

+++ sakai-11.4/samigo/samigo-api/src/java/org/sakaiproject/tool/assessment/shared/api/questionpool/QuestionPoolServiceAPI.java
@@ -74,7 +74,7 @@ public interface QuestionPoolServiceAPI
   /**
    * Get the size of a subpool.
    */
-  public long getSubPoolSize(Long poolId);
+  public int getSubPoolSize(Long poolId);
 
   /**
    * Checks to see if a pool has subpools
